splinter new squares component from board
