+++ src/css/Squares.css
@@ -1,0 +1,6 @@
+.square {
+    height: 100%;
+    width: 100%;
+    background: lightblue;
+    z-index: 0;
+}
